Create code in c++
Here’s the explanation:

sum: This is the sum of all absolute values.
zeros: If there are any zeros, they don't affect the result because flipping a zero doesn't change the sum.
negs % 2 == 0: If there’s an even number of negative values, it’s fine to leave the signs as is. However, if there’s an odd number of negative values, one negative number needs to be flipped to positive to maximize the sum. Flipping a negative number to positive means subtracting the smallest absolute value from the total sum (this is where the min_abs value comes into play).
2 * min_abs: This is the correction term when there's an odd number of negative numbers, where one negative value is flipped to positive to maximize the sum.